pyproject.toml
[tool.poetry]
name = "apite"
version = "0.1.0"
description = "API test automation tool"
authors = ["Prabakaran Santhanam"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
